+++ client/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added keyboard focus when navigating hierarchy ([#293](https://github.com/project-lux/lux-frontend/issues/293)).
+- Added access statement to works pages ([#364](https://github.com/project-lux/lux-frontend/issues/364)).
 
 ### Changed